refactored chat cards to reflect current user, non-friends, and preparing for friends

+++ src/components/chat/ChatCard.js
@@ -7,38 +7,51 @@ export const ChatCard = ({ message }) => {
     const { deleteMessage } = useContext(ChatContext)
     const history = useHistory()
 
-    return (
-        <Container className="message--container">
-            <Message className="message" floating>
-                <Modal
-                    trigger={<Header as="h3" className="message--poster">{message.user.username}</Header>}
-                    header="Test"
-                    content="Do you want to add Test as a friend?"
-                    actions={['No', { key: 'yes', content: 'Yes', positive: true }]}
-                />
+    if (message.userId === parseInt(localStorage.getItem("nutty_customer"))) {
 
-                <p className="message--content">{message.message}</p>
-                <p className="message--date">{message.date}</p>
+        // renders chat bubble for current user
+        return (
+            <Container className="message--container">
+                <Message className="message" floating style={{ backgroundColor: "#f7f7f7" }}>
+                    <Header as="h3" className="message--currentUser">{message.user.username}</Header>
 
-                <div className="messageBtns">
-                    {message.userId === parseInt(localStorage.getItem("nutty_customer")) ?
-                        <>
-                            <Button.Group>
-                                <Button icon className="editBtn" onClick={
-                                    () => history.push(`/messages/edit/${message.id}`)
-                                }><Icon name="edit" />
-                                </Button>
+                    <p className="message--content">{message.message}</p>
+                    <p className="message--date" style={{ fontSize: "x-small" }}>{message.date}</p>
 
-                                <Button icon className="deleteBtn" color="red" onClick={
-                                    () => deleteMessage(message.id)
-                                        .then(() => history.push(`/messages`))
-                                }><Icon name="trash" />
-                                </Button>
-                            </Button.Group>
-                        </>
-                        : null}
-                </div>
-            </Message>
-        </Container>
-    )
+                    <div className="messageBtns">
+                        <Button.Group>
+                            <Button icon className="editBtn" onClick={
+                                () => history.push(`/messages/edit/${message.id}`)
+                            }><Icon name="edit" />
+                            </Button>
+
+                            <Button icon className="deleteBtn" color="red" onClick={
+                                () => deleteMessage(message.id)
+                                    .then(() => history.push(`/messages`))
+                            }><Icon name="trash" />
+                            </Button>
+                        </Button.Group>
+                    </div>
+                </Message>
+            </Container>
+        )
+    } else {
+
+        // renders chat bubble for global, non-friend users
+        return (
+            <Container className="message--container">
+                <Message className="message" floating style={{ backgroundColor: "lightgreen" }}>
+                    <Modal
+                        trigger={<Header as="h3" className="message--notFriend">{message.user.username}</Header>}
+                        header="Test"
+                        content="Do you want to add Test as a friend?"
+                        actions={['No', { key: 'yes', content: 'Yes', positive: true }]}
+                    />
+
+                    <p className="message--content">{message.message}</p>
+                    <p className="message--date" style={{ fontSize: "x-small" }}>{message.date}</p>
+                </Message>
+            </Container>
+        )
+    }
 }
